requirements.txt
streamlit==1.32.0
pandas==2.1.4
plotly==5.22.0
python-docx==1.1.0
reportlab==4.0.9
supabase==2.4.0
httpx==0.24.1